ABP - Created new warehousing module
Created new module for containing the functionality to extract the
warehouse data for LFT.
